Refactoring and adding new audtorium repository
