+++ ApiApplication/Startup.cs
@@ -1,17 +1,13 @@
 using ApiApplication.Extensions;
 using ApiApplication.Services;
+using ApiApplication.Utils;
 using CinemaApplication.DAL;
-using CinemaApplication.DAL.Extensions;
 using CinemaApplication.Services.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using System;
-using System.Text.Json;
 
 namespace ApiApplication
 {
@@ -61,45 +57,5 @@ namespace ApiApplication
 
             SampleData.Initialize(app);
         }
-    }
-}
-
-//https://maximgorbatyuk.github.io/blog/development/2021-02-20-snake-case-and-asp-net-core/
-public class SnakeCaseNamingPolicy : JsonNamingPolicy
-{
-    public override string ConvertName(string name) => name.ToSnakeCase();
-}
-
-public static class JsonSerializationExtensions
-{
-    private static readonly SnakeCaseNamingStrategy _snakeCaseNamingStrategy
-        = new SnakeCaseNamingStrategy();
-
-    private static readonly JsonSerializerSettings _snakeCaseSettings = new JsonSerializerSettings
-    {
-        ContractResolver = new DefaultContractResolver
-        {
-            NamingStrategy = _snakeCaseNamingStrategy
-        }
-    };
-
-    public static string ToSnakeCase<T>(this T instance)
-    {
-        if (instance == null)
-        {
-            throw new ArgumentNullException(paramName: nameof(instance));
-        }
-
-        return JsonConvert.SerializeObject(instance, _snakeCaseSettings);
-    }
-
-    public static string ToSnakeCase(this string @string)
-    {
-        if (@string == null)
-        {
-            throw new ArgumentNullException(paramName: nameof(@string));
-        }
-
-        return _snakeCaseNamingStrategy.GetPropertyName(@string, false);
     }
 }
