+++ CinemaApplication.DAL/Abstractions/IMovieRepository.cs
